// upstream/extracted/apps/client/features/v1/frontend/shared/user/models.ts
export enum UserSettingKey {
	HAS_OPT_IN = "hasOptIn",
	IS_MEMORY_CONTEXT_ENABLED = "isMemoryContextEnabled",
	LANGUAGE = "language",
	TOAST_NOTIFICATIONS_ENABLED = "toastNotificationsEnabled",
	SPEECH_ENABLED = "speechEnabled",
	//IS_AB_TESTING_ENABLED = "isAbTestingEnabled", Unused for now
	IS_PROMPT_SUGGESTIONS_DISABLED = "isPromptSuggestionsDisabled",
	IS_CUSTOM_SYSTEM_PROMPT_ACTIVATED = "isCustomSystemPromptActivated",
	CUSTOM_SYSTEM_PROMPT = "customSystemPrompt",
	PROJECT_SECTION_ORDER = "projectSectionOrder",
	ARTIFACTS_ENABLED = "artifactsEnabled",
	AUTO_OPEN_ARTIFACTS = "autoOpenArtifacts",
	CODE_EXECUTION_ENABLED = "codeExecutionEnabled",
	CONVERSATION_STYLE = "conversationStyle",
	BUFFER_THREAD = "bufferThread",
	EMAIL_NOTIFICATIONS_ENABLED = "emailNotificationsEnabled",
}